Add basic skeleton for driver service
Added a simple WebSocket main loop with a couple of placeholder
functions to depict the basic workflow of messages.

The main idea is to provide a main loop with wrapper functions that run
in separate goRoutines.
These functions will be then handled synchronously in the service layer
and will provide the result with an output channel of type []byte

The basic error handling and process needs to be polished but the main
workflow seems solid for the moment

Signed-off-by: Oscar Moya <oscar.moya@alefedge.com>