Add location package
This package provides an interface to add/update driver locations,
remove driver locations and get drivers based on a radius in KMs.

The tool of choice has been Redis with its GeoAdd functionality. Since
this is a proof of concept and exploration architecture, no additional
keys have been added to mark or cache the driver locations to keep
things as simple as possible.

Since this functionality has a bit of logic, a proper test case has been
added.

A utility to add distance to long/lat pairs has been added to
simplify and automate as much as possible the test cases.

Last but not least, a simple dockerfile with Redis has been added to
make sure the tests can run easily

Signed-off-by: Oscar Moya <oscar.moya@alefedge.com>