Add basic authentication package
Added this package to handle the Creation and Validation of JWT tokens
for driver operations. The reason to need this is to make sure the
backend lifecycle is "secure" from the start and there are no messages
with timestamps or ids that can potentially be tweaked by malicious
users.

Even if for now the secrets are not managed and there is no
authentication per se, the development of operations can continue
without those while not breaking the security.

Signed-off-by: Oscar Moya <oscar.moya@alefedge.com>