Add CRUD API for Driver using PGX
Added basic PGX driver to handle CRUD operations with drivers.
There is still no relation between user and drivers and the model is
minimal but this is enough for the first implementation and the ability
to handle the lifecycle for drivers

Added basic test suite for all the operations and extended Makefile and
docker-compose to easily test this locally.

This API requires more work as it is a bit painful to extend

Signed-off-by: Oscar Moya <o.moya.gomez@gmail.com>