Refactor the Patterns for more code clarity
Corrected all the HTTP patterns to use a combination of service +
middleware where the services will interact most likely with other
packages like the repository, biller, queue, etc.

Aside from more extensibility, the DB interactions are way more solid
than the original implementation.

As part of this refactor, a SQL builder utility has been added. This
utility provides constructs to add create the Update/Query/Create
queries with a lot of extensibility using custom `db` struct tags +
scanners, this helps a lot in building the Queries sent to the DB with
the degree of flexibility needed and making sure there won't be any
vulnerability caused from the frontend.

Added basic tests in the ride_service so the workflow with the
notifications can be tested and validated.

Started to define a High level Interface for the ride service that
defines the status changes rather than the basic Crud.

Signed-off-by: Oscar Moya <o.moya.gomez@gmail.com>